Refactor markdown preview model files
